Support parsing and tagging lateral joins (#149)
Postgres has a LATERAL keyword that modifies one of the `from_item`s in a selection. It doesn't have to be an explicit join statement but is most often used as such. This adds support for parsing the LATERAL keyword in front of each from_item type that supports it.

See https://www.postgresql.org/docs/current/sql-select.html

Fixes #122